Abstracted bullet list component
