+++ app/components/BulletList/BulletList.d.ts
@@ -1,0 +1,6 @@
+import { ViewStyle } from 'react-native';
+
+export type BulletListProps = {
+  items: string[];
+  style?: StyleProp<ViewStyle>;
+};
