+++ app/components/index.ts
@@ -22,5 +22,6 @@ export { default as ImportFileModal } from './ImportFileModal/ImportFileModal';
 export { default as CredentialRequestHandler } from './CredentialRequestHandler/CredentialRequestHandler';
 export { default as ThemeProvider } from './ThemeProvider/ThemeProvider';
 export { default as GlobalErrorModal } from './GlobalErrorModal/GlobalErrorModal';
+export { default as BulletList } from './BulletList/BulletList';
 
 export * from './ImportFileModal/ImportFileModal.d';
