+++ app/screens/IssuerInfoScreen/IssuerInfoScreen.d.ts
@@ -3,7 +3,3 @@ export type { IssuerInfoScreenProps } from '../../navigation';
 export type IssuerInfoScreenParams = {
   issuerId: string;
 }
-
-export type BulletListProps = {
-  items: string[];
-};
